requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==20.1.0